third_party/blink/web_tests/fast/xmlhttprequest/xmlhttprequest-responseXML-invalid-xml-expected.txt

This tests the XMLHttpRequest responseXML attribute value when given invalid XML data.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xhr.readyState is 4
PASS xhr.responseXML is null
PASS successfullyParsed is true

TEST COMPLETE
